Change float128 to float64 for Apple silicon

+++ TP5/src/activation_functions.py
@@ -6,7 +6,7 @@ import numpy as np
 class Tanh(Activation):
     def __init__(self, beta: float = 0.4):
         def tanh(x, beta: float):
-            return np.tanh(np.float128(x) * beta)
+            return np.tanh(np.float64(x) * beta)
 
         def tanh_prime(x, beta: float):
             return 1 - tanh(x, beta) ** 2
@@ -18,7 +18,7 @@ class Tanh(Activation):
 class ReLU(Activation):
     def __init__(self, beta: float = 1.0):
         def relu(x, beta: float):
-            return np.maximum(0, np.float128(x))
+            return np.maximum(0, np.float64(x))
 
         def relu_prime(x, beta: float):
             return np.where(x > 0, 1, 0)
@@ -29,7 +29,7 @@ class ReLU(Activation):
 class Logistic(Activation):
     def __init__(self, beta: float = 0.4):
         def logistic(x, beta: float):
-            return 1 / (1 + np.exp(-2 * beta * np.float128(x)))
+            return 1 / (1 + np.exp(-2 * beta * np.float64(x)))
 
         def logistic_prime(x, beta: float):
             return 2 * beta * logistic(x, beta) * (1 - logistic(x, beta))
@@ -40,7 +40,7 @@ class Logistic(Activation):
 class Sigmoid(Activation):
     def __init__(self, beta: float = 1.0):
         def sigmoid(x, beta: float):
-            return 1 / (1 + np.exp(-beta * np.float128(x)))
+            return 1 / (1 + np.exp(-beta * np.float64(x)))
 
         def sigmoid_prime(x, beta: float):
             sig = sigmoid(x, beta)
@@ -52,10 +52,10 @@ class Sigmoid(Activation):
 class Linear(Activation):
     def __init__(self, beta: float = 1.0):
         def linear(x, beta: float):
-            return beta * np.float128(x)
+            return beta * np.float64(x)
 
         def linear_prime(x, beta: float):
-            return beta * np.ones_like(np.float128(x))
+            return beta * np.ones_like(np.float64(x))
 
         super().__init__(linear, linear_prime, beta)
